Add files and progress on project.
This is an updated version of the pomodoro repository that I'm not sure how to delete. Anyway in this project, I figured out how to get a working notepad in, update the timer based on changed input values, aaand have a good solution for the side panels(accordions) for mobile.